refactor: Extract EvalState::addCallDepth
