+++ src/libexpr/eval.hh
@@ -42,6 +42,21 @@ namespace eval_cache {
 }
 
 /**
+ * Increments a count on construction and decrements on destruction.
+ */
+class CallDepth {
+  size_t & count;
+
+public:
+  CallDepth(size_t & count) : count(count) {
+    ++count;
+  }
+  ~CallDepth() {
+    --count;
+  }
+};
+
+/**
  * Function that implements a primop.
  */
 using PrimOpFun = void(EvalState & state, const PosIdx pos, Value * * args, Value & v);
@@ -648,6 +663,11 @@ private:
     size_t callDepth = 0;
 
 public:
+
+    /**
+     * Check that the call depth is within limits, and increment it, until the returned object is destroyed.
+     */
+    inline CallDepth addCallDepth(const PosIdx pos);
 
     /**
      * Do a deep equality test between two values.  That is, list
